pyproject.toml
[tool]
[tool.poetry]
authors = ["@spacesanjeet"]
name = "root"
version = "0.0.0"
[tool.poetry.dependencies]
"discord.py" = "*"
flask = "*"
python = "^3.7"
replit = "*"
chucknorris = "^2.8"
xkcd = "^2.4"
